Extract footer actions.
